+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Node IDs in matches don't have the `salt:/` prefix anymore
 
+### Added
+
+- Add non-tokenized primary text segments as special nodes with the node type "white-space-token" when importing relANNIS.
+  These are connected with edges in the Ordering/annis/relannis.text component to the other token and allow to re-construct the
+  original relANNIS primary text.
+
+## [0.28.0] - 2020-07-02
+
 ### Addded
 
 - Web Service with REST API for the corpus storage